+++ packages/castle/src/exceptions/index.ts
@@ -1,3 +1,3 @@
 export { CastleException } from './castle_exception.js'
-export { parseError } from './parse.js'
+export { parseError, isCastleException, isException } from './utils.js'
 export * from './http_exceptions.js'