+++ playground/backend/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @folie/playground-backend
 
+## 0.0.22
+
+### Patch Changes
+
+- Updated dependencies
+  - @folie/castle@0.0.32
+
 ## 0.0.21
 
 ### Patch Changes